requirements.txt
google-cloud-storage==2.10.0
google-auth==2.21.0
whisper==1.1.7
nltk==3.8.1
regex==2023.6.3
concurrent-log-handler==0.9.24
python-dotenv==1.0.0
tqdm==4.65.0
google-api-python-client==2.97.0
Flask==2.3.2

